Slider1.0Refact
