+++ src/components/slider/slider-cards.tsx
@@ -1,0 +1,22 @@
+import React from "react";
+
+
+// interface SliderCardProps {
+//     header: string;
+//     desc: string;
+//     image: ReactElement;
+// }
+
+// const SliderCard1: React.FC<SliderCardProps> = ({ header, desc, image }) => {
+//     return (
+//         <div className="container">
+//             {image}
+//             <h1>{header}</h1>
+//             <p>{desc}</p>
+//         </div>
+//     );
+// };
+
+// export {
+//     SliderCard1
+// };
