+++ src/components/slider/slider.tsx
@@ -1,68 +1,69 @@
-// import React, { Component } from "react";
-// import Slider from "react-slick";
-// import { BurgerSVG } from "../../icons/burger";
-// import FruitSVG from "../../icons/strawb";
-// import "react-awesome-slider/src/core/styles.scss";
+import React, { ReactElement } from "react";
+import Slider from "react-slick";
+// import { SliderCard1 } from "./slider-cards";
+
+import StartPageSecImage from "./images/st-pg-sec-img";
+import StartPageFirstImage from "./images/st-pg-first-img";
+import StartPageThirImage from "./images/st-pg-thir-img";
 
 
-// export default class SimpleSlider extends Component {
-//     render() {
-//         const settings = {
-//             dots: true,
-//             infinite: true,
-//             speed: 500,
-//             slidesToShow: 1,
-//             slidesToScroll: 1
-//         };
-//         return (
-//             <div>
-//                 <h2> Single Item</h2>
-//                 <Slider {...settings}>
-//                     <div>
-//                         <h3>1</h3>
-//                     </div>
-//                     <div>
-//                         <h3>2</h3>
-//                     </div>
-//                     <div>
-//                         <h3>3</h3>
-//                     </div>
-//                     <div>
-//                         <h3>4</h3>
-//                     </div>
-//                     <div>
-//                         <h3>5</h3>
-//                     </div>
-//                     <div>
-//                         <h3>6</h3>
-//                     </div>
-//                 </Slider>
-//             </div>
-//         );
-//     }
-// }
+interface SliderCardProps {
+    header: string;
+    desc: string;
+    image: ReactElement;
+}
 
-// Import Swiper React components
-import React from 'react';
-import { Swiper, SwiperSlide } from 'swiper/react';
+const SliderCard1: React.FC<SliderCardProps> = ({ header, desc, image }) => {
+    return (
+        <div className="container">
+            {image}
+            <h1>{header}</h1>
+            <p>{desc}</p>
+        </div>
+    );
+};
 
-// Import Swiper styles
-// import 'swiper/css';
+export {
+    SliderCard1
+};
 
-// const Swipper = () => {
-//     return (
-//         <Swiper
-//             spaceBetween={50}
-//             slidesPerView={3}
-//             onSlideChange={() => console.log('slide change')}
-//             onSwiper={(swiper) => console.log(swiper)}
-//         >
-//             <SwiperSlide>Slide 1</SwiperSlide>
-//             <SwiperSlide>Slide 2</SwiperSlide>
-//             <SwiperSlide>Slide 3</SwiperSlide>
-//             <SwiperSlide>Slide 4</SwiperSlide>
-//             ...
-//         </Swiper>
-//     );
-// };
-// export default Swipper;
+class ReactSlickDemo extends React.Component {
+    render() {
+        var settings = {
+            dots: true,
+
+        };
+        return (
+            <div className="container">
+                <Slider   {...settings}>
+                    <div>
+                        <SliderCard1 header="Eat Healthy" image={<StartPageFirstImage />}
+                            desc="Maintaining good health should be the primary focus of everyone.">
+
+                        </SliderCard1>
+                    </div>
+                    <div>
+                        <SliderCard1 header="Healthy Recipes" image={<StartPageSecImage />}
+                            desc="Browse thousands of healthy recipes from all over the world.">
+
+                        </SliderCard1>
+                    </div>
+                    <div>
+                        <SliderCard1 header="Track Your Health" image={<StartPageThirImage />}
+                            desc="With amazing inbuilt tools you can track your    progress.">
+
+                        </SliderCard1>
+                    </div>
+                    <div>
+                        <StartPageThirImage />
+                        <h1>Track Your Health</h1>
+                        <p>With amazing inbuilt tools you can track your progress.</p>
+                    </div>
+                </Slider>
+
+            </div>
+        );
+    }
+}
+
+export default ReactSlickDemo;
