+++ src/components/slider/test-slider/slider.scss
