+++ src/components/slider/test-slider/slider.tsx
@@ -1,0 +1,68 @@
+// import React, { Component } from "react";
+// import Slider from "react-slick";
+// import { BurgerSVG } from "../../icons/burger";
+// import FruitSVG from "../../icons/strawb";
+// import "react-awesome-slider/src/core/styles.scss";
+
+
+// export default class SimpleSlider extends Component {
+//     render() {
+//         const settings = {
+//             dots: true,
+//             infinite: true,
+//             speed: 500,
+//             slidesToShow: 1,
+//             slidesToScroll: 1
+//         };
+//         return (
+//             <div>
+//                 <h2> Single Item</h2>
+//                 <Slider {...settings}>
+//                     <div>
+//                         <h3>1</h3>
+//                     </div>
+//                     <div>
+//                         <h3>2</h3>
+//                     </div>
+//                     <div>
+//                         <h3>3</h3>
+//                     </div>
+//                     <div>
+//                         <h3>4</h3>
+//                     </div>
+//                     <div>
+//                         <h3>5</h3>
+//                     </div>
+//                     <div>
+//                         <h3>6</h3>
+//                     </div>
+//                 </Slider>
+//             </div>
+//         );
+//     }
+// }
+
+// Import Swiper React components
+import React from 'react';
+import { Swiper, SwiperSlide } from 'swiper/react';
+
+// Import Swiper styles
+// import 'swiper/css';
+
+// const Swipper = () => {
+//     return (
+//         <Swiper
+//             spaceBetween={50}
+//             slidesPerView={3}
+//             onSlideChange={() => console.log('slide change')}
+//             onSwiper={(swiper) => console.log(swiper)}
+//         >
+//             <SwiperSlide>Slide 1</SwiperSlide>
+//             <SwiperSlide>Slide 2</SwiperSlide>
+//             <SwiperSlide>Slide 3</SwiperSlide>
+//             <SwiperSlide>Slide 4</SwiperSlide>
+//             ...
+//         </Swiper>
+//     );
+// };
+// export default Swipper;
