+++ src/global.scss
@@ -11,18 +11,18 @@ body {
 }
 
 .container {
-	padding: -5px;
+	// padding: px;
 	background: beige;
 	margin-top: 2rem;
 	border-radius: 40px;
 	font-family: "Signika";
 	font-weight: bold;
 	max-width: 300px;
-	max-height: 650px;
+	max-height: 700px;
 	display: flex;
-    flex-direction: column-reverse;
-  	text-align: center;
-  	justify-content: flex-end;
+    flex-direction: column;
+	justify-content: start;
+	text-align: center;
 }
