+++ src/index.tsx
@@ -8,7 +8,7 @@ import { BurgerSVG } from "./icons/burger";
 import FruitSVG from "./icons/strawb";
 import { SquareBut } from "./components/buttons/square-button";
 import RectangleButton, { RectBut } from "./components/buttons/rectangle-button";
-import ReactSlickDemo from "./components/slider/test-slider";
+import ReactSlickDemo from "./components/slider/slider";
 import ImageSlider from "./components/slider/test-slider/ImageSlider";
 // import SimpleSlider from "./components/slider/slider";
